move functions into helper package
