+++ main.go
@@ -16,15 +16,15 @@ import (
 	"text/template"
 	"time"
 
+	"github.com/libops/captcha-protect/internal/helper"
+	plog "github.com/libops/captcha-protect/internal/log"
 	"github.com/libops/captcha-protect/internal/state"
-	"github.com/libops/captcha-protect/internal/tmpl"
+
 	lru "github.com/patrickmn/go-cache"
 )
 
 var (
-	lookupAddrFunc = net.LookupAddr
-	lookupIPFunc   = net.LookupIP
-	log            *slog.Logger
+	log *slog.Logger
 )
 
 type Config struct {
@@ -107,18 +107,7 @@ func New(ctx context.Context, next http.Handler, config *Config, name string) (h
 }
 
 func NewCaptchaProtect(ctx context.Context, next http.Handler, config *Config, name string) (*CaptchaProtect, error) {
-	var logLevel slog.LevelVar
-	logLevel.Set(slog.LevelInfo)
-	handler := slog.NewTextHandler(os.Stdout, &slog.HandlerOptions{
-		Level: &logLevel,
-	})
-	log = slog.New(handler)
-
-	level, err := ParseLogLevel(config.LogLevel)
-	if err != nil {
-		log.Warn("Unknown log level", "err", err)
-	}
-	logLevel.Set(level)
+	log = plog.New(config.LogLevel)
 
 	expiration := time.Duration(config.Window) * time.Second
 	log.Debug("Captcha config", "config", config)
@@ -126,6 +115,7 @@ func NewCaptchaProtect(ctx context.Context, next http.Handler, config *Config, n
 	if len(config.ProtectRoutes) == 0 && config.Mode != "suffix" {
 		return nil, fmt.Errorf("you must protect at least one route with the protectRoutes config value. / will cover your entire site")
 	}
+
 	protectRoutesRegex := []*regexp.Regexp{}
 	excludeRoutesRegex := []*regexp.Regexp{}
 	if config.Mode == "regex" {
@@ -159,18 +149,18 @@ func NewCaptchaProtect(ctx context.Context, next http.Handler, config *Config, n
 	}
 	config.ParseHttpMethods()
 
-	var t *template.Template
+	var tmpl *template.Template
 	if _, err := os.Stat(config.ChallengeTmpl); os.IsNotExist(err) {
 		log.Warn("Unable to find template file. Using default template.", "challengeTmpl", config.ChallengeTmpl)
-		ts := tmpl.GetDefaultTmpl()
-		t, err = template.New("challenge").Parse(ts)
+		ts := helper.GetDefaultTmpl()
+		tmpl, err = template.New("challenge").Parse(ts)
 		if err != nil {
 			return nil, fmt.Errorf("unable to parse challenge template: %v", err)
 		}
 	} else if err != nil {
 		return nil, fmt.Errorf("error checking for template file %s: %v", config.ChallengeTmpl, err)
 	} else {
-		t, err = template.ParseFiles(config.ChallengeTmpl)
+		tmpl, err = template.ParseFiles(config.ChallengeTmpl)
 		if err != nil {
 			return nil, fmt.Errorf("unable to parse challenge template file %s: %v", config.ChallengeTmpl, err)
 		}
@@ -191,7 +181,7 @@ func NewCaptchaProtect(ctx context.Context, next http.Handler, config *Config, n
 	}
 	exemptIps = append(exemptIps, config.ExemptIPs...)
 	for _, ip := range exemptIps {
-		parsedIp, err := ParseCIDR(ip)
+		parsedIp, err := helper.ParseCIDR(ip)
 		if err != nil {
 			return nil, fmt.Errorf("error parsing cidr %s: %v", ip, err)
 		}
@@ -206,12 +196,12 @@ func NewCaptchaProtect(ctx context.Context, next http.Handler, config *Config, n
 		botCache:           lru.New(expiration, 1*time.Hour),
 		verifiedCache:      lru.New(expiration, 1*time.Hour),
 		exemptIps:          ips,
-		tmpl:               t,
+		tmpl:               tmpl,
 		protectRoutesRegex: protectRoutesRegex,
 		excludeRoutesRegex: excludeRoutesRegex,
 	}
 
-	err = bc.SetIpv4Mask(config.IPv4SubnetMask)
+	err := bc.SetIpv4Mask(config.IPv4SubnetMask)
 	if err != nil {
 		return nil, err
 	}
@@ -384,7 +374,7 @@ func (bc *CaptchaProtect) verifyChallengePage(rw http.ResponseWriter, req *http.
 
 func (bc *CaptchaProtect) serveStatsPage(rw http.ResponseWriter, ip string) {
 	// only allow excluded IPs from viewing
-	if !IsIpExcluded(ip, bc.exemptIps) {
+	if !helper.IsIpExcluded(ip, bc.exemptIps) {
 		http.Error(rw, "Forbidden", http.StatusForbidden)
 		return
 	}
@@ -418,7 +408,7 @@ func (bc *CaptchaProtect) shouldApply(req *http.Request, clientIP string) bool {
 		return false
 	}
 
-	if IsIpExcluded(clientIP, bc.exemptIps) {
+	if helper.IsIpExcluded(clientIP, bc.exemptIps) {
 		return false
 	}
 
@@ -536,17 +526,6 @@ protected:
 	return false
 }
 
-func IsIpExcluded(clientIP string, exemptIps []*net.IPNet) bool {
-	ip := net.ParseIP(clientIP)
-	for _, block := range exemptIps {
-		if block.Contains(ip) {
-			return true
-		}
-	}
-
-	return false
-}
-
 func (bc *CaptchaProtect) trippedRateLimit(ip string) bool {
 	v, ok := bc.rateCache.Get(ip)
 	if !ok {
@@ -576,7 +555,7 @@ func (bc *CaptchaProtect) getClientIP(req *http.Request) (string, string) {
 		ip = ""
 		for i := len(components) - 1; i >= 0; i-- {
 			_ip := strings.TrimSpace(components[i])
-			if IsIpExcluded(_ip, bc.exemptIps) {
+			if helper.IsIpExcluded(_ip, bc.exemptIps) {
 				continue
 			}
 			if depth == 0 {
@@ -656,74 +635,13 @@ func (bc *CaptchaProtect) isGoodBot(req *http.Request, clientIP string) bool {
 		return bot.(bool)
 	}
 
-	v := IsIpGoodBot(clientIP, bc.config.GoodBots)
+	v := helper.IsIpGoodBot(clientIP, bc.config.GoodBots)
 	bc.botCache.Set(clientIP, v, lru.DefaultExpiration)
 	return v
 }
 
-func IsIpGoodBot(clientIP string, goodBots []string) bool {
-	if len(goodBots) == 0 {
-		return false
-	}
-
-	// lookup the hostname for a given IP
-	hostname, err := lookupAddrFunc(clientIP)
-	if err != nil || len(hostname) == 0 {
-		return false
-	}
-
-	// then nslookup that hostname to avoid spoofing
-	resolvedIP, err := lookupIPFunc(hostname[0])
-	if err != nil || len(resolvedIP) == 0 || resolvedIP[0].String() != clientIP {
-		return false
-	}
-
-	// get the sld
-	// will be like 194.114.135.34.bc.googleusercontent.com.
-	// notice the trailing period
-	parts := strings.Split(hostname[0], ".")
-	l := len(parts)
-	if l < 3 {
-		return false
-	}
-	tld := parts[l-2]
-	domain := parts[l-3] + "." + tld
-
-	for _, bot := range goodBots {
-		if domain == bot {
-			return true
-		}
-	}
-
-	return false
-}
-
 func (bc *CaptchaProtect) SetExemptIps(exemptIps []*net.IPNet) {
 	bc.exemptIps = exemptIps
-}
-
-func ParseCIDR(cidr string) (*net.IPNet, error) {
-	_, ipNet, err := net.ParseCIDR(cidr)
-	if err != nil {
-		return nil, err
-	}
-	return ipNet, nil
-}
-
-// Map string to slog.Level
-func ParseLogLevel(level string) (slog.Level, error) {
-	switch strings.ToUpper(level) {
-	case "DEBUG":
-		return slog.LevelDebug, nil
-	case "INFO":
-		return slog.LevelInfo, nil
-	case "WARNING", "WARN":
-		return slog.LevelWarn, nil
-	case "ERROR":
-		return slog.LevelError, nil
-	default:
-		return slog.LevelInfo, fmt.Errorf("unknown logl level %s", level)
-	}
 }
 
 // log a warning if protected methods contains an invalid method
